# apis/bijux-canon-index/v1/pinned_openapi.json
{
  "components": {
    "schemas": {
      "ArtifactResponse": {
        "additionalProperties": false,
        "description": "Represents artifact response.",
        "properties": {
          "artifact_id": {
            "description": "Identifier of the execution artifact materialized by the API.",
            "title": "Artifact Id",
            "type": "string"
          },
          "execution_contract": {
            "description": "Execution contract currently attached to the artifact.",
            "title": "Execution Contract",
            "type": "string"
          },
          "execution_contract_status": {
            "description": "Human-readable stability marker for the chosen contract.",
            "title": "Execution Contract Status",
            "type": "string"
          },
          "replayable": {
            "description": "Whether the artifact can be replayed under its declared contract.",
            "title": "Replayable",
            "type": "boolean"
          }
        },
        "required": [
          "artifact_id",
          "execution_contract",
          "execution_contract_status",
          "replayable"
        ],
        "title": "ArtifactResponse",
        "type": "object"
      },
...
      "CreateRequest": {
        "additionalProperties": false,
        "description": "Represents create request.",
        "properties": {
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateRequest",
        "type": "object"
      },
      "CreateResponse": {
        "additionalProperties": false,
        "description": "Represents create response.",
        "properties": {
          "name": {
            "description": "Logical corpus name reserved by the API.",
            "title": "Name",
            "type": "string"
          },
          "status": {
            "description": "Lifecycle state returned after the create request.",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "status"
        ],
        "title": "CreateResponse",
        "type": "object"
      },
...
      "ExecutionArtifactRequest": {
        "additionalProperties": false,
        "description": "Represents execution artifact request.",
        "properties": {
          "execution_contract": {
            "$ref": "#/components/schemas/ExecutionContract"
          },
          "index_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index Mode"
          },
          "vector_store": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Store"
          },
          "vector_store_options": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Store Options"
          },
          "vector_store_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Store Uri"
          }
        },
        "required": [
          "execution_contract"
        ],
        "title": "ExecutionArtifactRequest",
        "type": "object"
      },
...
      "ExecutionContract": {
        "description": "Enumeration of execution contract.",
        "enum": [
          "deterministic",
          "non_deterministic"
        ],
        "title": "ExecutionContract",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestRequest": {
        "additionalProperties": false,
        "description": "Represents ingest request.",
        "properties": {
          "cache_embeddings": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Embeddings"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "documents": {
            "items": {
              "type": "string"
            },
            "title": "Documents",
            "type": "array"
          },
          "embed_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embed Model"
          },
          "embed_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embed Provider"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "vector_store": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Store"
          },
          "vector_store_options": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Store Options"
          },
          "vector_store_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Store Uri"
          },
          "vectors": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vectors"
          }
        },
        "required": [
          "documents"
        ],
        "title": "IngestRequest",
        "type": "object"
      },
      "IngestResponse": {
        "additionalProperties": false,
        "description": "Represents ingest response.",
        "properties": {
          "correlation_id": {
            "description": "Correlation identifier attached to the ingest transaction.",
            "title": "Correlation Id",
            "type": "string"
          },
          "ingested": {
            "description": "Number of documents accepted into the backend during this call.",
            "minimum": 0.0,
            "title": "Ingested",
            "type": "integer"
          }
        },
        "required": [
          "ingested",
          "correlation_id"
        ],
        "title": "IngestResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Bijux",
      "url": "https://github.com/bijux/bijux-canon"
    },
    "description": "The bijux-canon-index HTTP API exposes the v1 surface for deterministic and bounded vector execution. It separates capability discovery, artifact materialization, and query execution so operators can review the contract they are invoking before they ship it.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "summary": "Contract-driven vector execution and replayable retrieval.",
    "title": "bijux-canon-index API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/artifact": {
      "post": {
        "description": "Freeze the current corpus and vector state into an execution artifact with a declared contract so later executions can be audited and replayed.",
        "operationId": "materializeExecutionArtifact",
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "artifact": {
                  "value": {
                    "execution_contract": "deterministic"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ExecutionArtifactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "artifact": {
                    "value": {
                      "artifact_id": "art-1",
                      "execution_contract": "deterministic",
                      "execution_contract_status": "stable",
                      "replayable": true
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            },
            "description": "Execution artifact materialized successfully."
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "refusal": {
                    "value": {
                      "error": {
                        "message": "[INV-000] Deterministic execution requires a deterministic vector store",
                        "reason": "determinism_violation",
                        "remediation": "Use deterministic inputs or switch to non_deterministic contract with declared randomness."
                      }
                    }
                  }
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Materialize an execution artifact",
        "tags": [
          "Materialization"
        ]
      }
    },
...
    "/create": {
      "post": {
        "description": "Create a logical corpus entry before ingestion. The endpoint is intentionally small so teams can separate naming and governance from later content writes.",
        "operationId": "createCorpus",
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "create": {
                  "value": {
                    "name": "my-corpus"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "created": {
                    "value": {
                      "name": "my-corpus",
                      "status": "created"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/CreateResponse"
                }
              }
            },
            "description": "Corpus name reserved successfully."
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "refusal": {
                    "value": {
                      "error": {
                        "message": "[INV-000] Deterministic execution requires a deterministic vector store",
                        "reason": "determinism_violation",
                        "remediation": "Use deterministic inputs or switch to non_deterministic contract with declared randomness."
                      }
                    }
                  }
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reserve a logical corpus name",
        "tags": [
          "Materialization"
        ]
      }
    },
...
    "/ingest": {
      "post": {
        "description": "Persist documents and either supplied vectors or embedding-derived vectors, then return the ingest count and resolved correlation identifier.",
        "operationId": "ingestDocuments",
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "ingest": {
                  "value": {
                    "documents": [
                      "hello"
                    ],
                    "vector_store": "faiss",
                    "vector_store_uri": "index.faiss",
                    "vectors": [
                      [
                        0.0,
                        1.0,
                        0.0
                      ]
                    ]
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "ingested": {
                    "value": {
                      "correlation_id": "req-example",
                      "ingested": 1
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Documents ingested successfully."
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "refusal": {
                    "value": {
                      "error": {
                        "message": "[INV-000] Deterministic execution requires a deterministic vector store",
                        "reason": "determinism_violation",
                        "remediation": "Use deterministic inputs or switch to non_deterministic contract with declared randomness."
                      }
                    }
                  }
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest documents and vectors",
        "tags": [
          "Materialization"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}